Move the LayoutType enum to the data class
