+++ browser-api/src/main/java/com/duckduckgo/app/tabs/model/TabSwitcherData.kt
@@ -20,10 +20,15 @@ data class TabSwitcherData(
     val userState: UserState,
     val wasAnnouncementDismissed: Boolean,
     val announcementDisplayCount: Int,
+    val layoutType: LayoutType,
 ) {
     enum class UserState {
         NEW,
         EXISTING,
         UNKNOWN,
+    }
+
+    enum class LayoutType {
+        GRID, LIST
     }
 }
